removed logos of non consortium members
from the box listing the consortium members. we will need to find another place for them, but they should not be listed alongside the paying members (that is a benefit of being a paying member). we will also need to discuss what criteria we use for showing a logo of a non-member partner. I.e why does KL gets its logo shown and why does stanford not get the logo?